upartner/requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
Django==1.9
django-celery==3.1.17
django-filter==0.11.0
djangorestframework==3.3.1
kombu==3.0.30
Markdown==2.6.5
psycopg2==2.6.1
pytz==2015.7
